pyproject.toml
[project]
name = "nibe750"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pymongo>=4.11.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
